# Conflicts:
#	src/main/java/com/spring/bittlebittle/user/dao/UserDaoImpl.java
#	src/main/java/com/spring/bittlebittle/user/service/UserServiceImpl.java

+++ src/main/java/com/spring/bittlebittle/tag/controller/AdminTagController.java
@@ -1,8 +1,11 @@
 package com.spring.bittlebittle.tag.controller;
 
 import java.util.List;
-import com.spring.bittlebittle.tag.service.TagService;
-import com.spring.bittlebittle.tag.vo.Tag;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,6 +13,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.spring.bittlebittle.tag.service.TagService;
+import com.spring.bittlebittle.tag.vo.Tag;
+import com.spring.bittlebittle.tag.vo.TagType;
+import com.spring.bittlebittle.user.vo.UserJwt;
+import com.spring.bittlebittle.utils.JwtUtil;
 
 
 
@@ -19,10 +28,15 @@ public class AdminTagController {
 
 	@Autowired
 	private TagService tservice;
+	@Autowired
+	private JwtUtil jwtUtil;
+	
+	Logger log = LogManager.getLogger("case3");
 	
 	// 태그 창 들어갔을 때 (확인완료)
 	@GetMapping
-	public List<Tag>  getTags(){
+
+	public List<Tag>  getTags() {
 
 		List<Tag> tagList = tservice.getAllTags();
 		
@@ -30,31 +44,78 @@ public class AdminTagController {
 	}
 	// 작성 (확인완료)
 	@PostMapping
-	public List<Tag> addTagType(@RequestBody Tag tag){
+	public List<Tag> addTag(@RequestBody Tag tag, HttpServletRequest request){
 								// keyTypeNo, tagName
 		
-		List<Tag> tagList = tservice.addTag(tag);
-		
-		return tagList;
+		String token = jwtUtil.resolveAccessToken(request);
+		String refreshTokenIdx = jwtUtil.resolveRefreshToken(request);
+		log.debug(token);
+		log.debug(refreshTokenIdx);
+		if (jwtUtil.validateToken(token, UserJwt.builder()
+				.userJwtIdx(refreshTokenIdx)
+				.build())) {
+			
+			List<Tag> tagList = tservice.addTag(tag);
+			
+			return tagList;
+			
+		} else {
+			
+			List<Tag> tagList = tservice.getAllTags();
+			
+			return tagList;
+		}
+	
 	}
 	
 	// 수정 (확인완료)
 	@PostMapping(value="/set-data")
-	public List<Tag> editTag(@RequestBody Tag tag){
+	public List<Tag> editTag(@RequestBody Tag tag, HttpServletRequest request){
 										// TagNo, TagType, keyTypeNo
 		
-		List<Tag> tagList = tservice.editTag(tag);
+		String token = jwtUtil.resolveAccessToken(request);
+		String refreshTokenIdx = jwtUtil.resolveRefreshToken(request);
+		log.debug(token);
+		log.debug(refreshTokenIdx);
+		if (jwtUtil.validateToken(token, UserJwt.builder()
+				.userJwtIdx(refreshTokenIdx)
+				.build())) {
+			
+			List<Tag> tagList = tservice.editTag(tag);
+			
+			return tagList;
+			
+		} else {
+			
+			List<Tag> tagList = tservice.getAllTags();
+			
+			return tagList;
+		}
 		
-		return tagList;
 	}
 	
 	// 삭제 (확인완료)
 	@GetMapping(value="/{tagNo}/deletion")
-	public List<Tag> removeTagType(@PathVariable int tagNo){
+	public List<Tag> removeTagType(@PathVariable int tagNo, HttpServletRequest request){
 		
-		List<Tag> tagList = tservice.removeTag(tagNo);
-		
-		return tagList;
+		String token = jwtUtil.resolveAccessToken(request);
+		String refreshTokenIdx = jwtUtil.resolveRefreshToken(request);
+		log.debug(token);
+		log.debug(refreshTokenIdx);
+		if (jwtUtil.validateToken(token, UserJwt.builder()
+				.userJwtIdx(refreshTokenIdx)
+				.build())) {
+			
+			List<Tag> tagList = tservice.removeTag(tagNo);
+			
+			return tagList;
+			
+		} else {
+			
+			List<Tag> tagList = tservice.getAllTags();
+			
+			return tagList;
+		}
 		
 	}
